- [Mac] Converted all binaries and prebuilt libraries to x86_64 (min 10.6) only - [Mac] Removed Xcode3 project and Xcode3 build scripts - [Mac] Converted sources and projects to ARC and Objective C 2.0 syntax - [Mac] Silenced most of the compiler and static analyser warnings in OUR code - [Mac] Removed the compatibility module proximus - no longer needed - [Mac] Modernized standalone application full screen behaviour. - [Mac] Converted installer scripts to use pkgbuild + productbuild Apple tools. No more PackageMaker - [Mac] Removed PackageMaker installer source files - [Mac] Flattened installer and signing with "Installer" DeveloperID certificate - [Mac] Updated Sparkle framework to version 1.8 - [Mac] Updated ffmpeg library to version 2.4
- Sealed hopefully all remaining obvious memory leaks
- Updated tinyxml library to 2.6.2

+++ client_generic/DisplayOutput/OpenGL/FontGL.cpp
@@ -17,7 +17,7 @@
 
 #if defined( WIN32 ) || defined( __LITTLE_ENDIAN__ )
 	#define SWAP_LONG_(x) ((x << 24) | ((x << 8) & 0x00FF0000) |  ((x >> 8) & 0x0000FF00) | (x >> 24))
-	#define SWAP_SHORT_(x) ((x << 8) | (x >> 8))
+	#define SWAP_SHORT_(x) ((unsigned short)((unsigned short)(x) << 8) | (unsigned short)((unsigned short)(x) >> 8))
 #else
 	#define SWAP_LONG_(x) x
 	#define SWAP_SHORT_(x) x
@@ -167,7 +167,7 @@ fp4 CFontGL::StringWidth(const std::string& str) const
 {
 	fp4 total = 0.0;
 	for (uint32 i = 0; i != str.size(); ++i)
-		total += CharWidth(str[i]);
+		total += CharWidth(static_cast<uint8>(str[i]));
 	return total;
 }
